Fixes #8979: fold epic sweep into monitor
