+++ docs/arch/coverage_matrix.md
@@ -77,7 +77,6 @@ None — every loop has substantive (non-roll-call) coverage when an ADR mention
 | `EdgeProposerLoop` | ✅ [0058](../adr/0058-edge-proposer-loop.md), [0060](../adr/0060-atlas-graph-view-and-provenance.md), [0062](../adr/0062-entry-evidence-loop.md) | ❌ [bd:advisor-u3m] | ❌ [bd:advisor-9i3] | ✅ (caretaking/caretaker loop) | ✅ `test_edge_proposer_loop.py` | ❌ [bd:advisor-2kq] | ❌ [bd:advisor-vwh] |
 | `EntryEvidenceLoop` | ✅ [0062](../adr/0062-entry-evidence-loop.md) | ❌ [bd:advisor-byh] | ❌ [bd:advisor-6ru] | ✅ (caretaking/caretaker loop) | ✅ `test_entry_evidence_loop.py` | ⚠️ [bd:advisor-4dj] | ❌ [bd:advisor-7m5] |
 | `EpicMonitorLoop` | ❌ [bd:advisor-o9d] | ❌ [bd:advisor-c88] | ❌ [bd:advisor-goo] | ✅ (caretaking/caretaker loop) | ✅ `test_epic_monitor_loop.py` | ✅ in catalog | ❌ [bd:advisor-lgd] |
-| `EpicSweeperLoop` | ❌ [bd:advisor-0zt] | ❌ [bd:advisor-j43] | ❌ [bd:advisor-8sg] | ✅ (caretaking/caretaker loop) | ✅ `test_epic_sweeper_loop.py` | ⚠️ in catalog [bd:advisor-4m0] | ❌ [bd:advisor-538] |
 | `FakeCoverageAuditorLoop` | ✅ [0045](../adr/0045-trust-architecture-hardening.md), [0056](../adr/0056-adr-touchpoint-gate-to-caretaker-loop.md) | ❌ [bd:advisor-t3h] | ❌ [bd:advisor-aqt] | ❌ [bd:advisor-15g] | ✅ `test_fake_coverage_auditor_loop.py` | ✅ in catalog | ❌ [bd:advisor-ln3] |
 | `FlakeTrackerLoop` | ✅ [0045](../adr/0045-trust-architecture-hardening.md), [0056](../adr/0056-adr-touchpoint-gate-to-caretaker-loop.md) | ❌ [bd:advisor-ifr] | ❌ [bd:advisor-c6x] | ❌ [bd:advisor-7pg] | ✅ `test_flake_tracker_loop.py` | ✅ in catalog | ❌ [bd:advisor-r8i] |
 | `GitHubCacheLoop` | ❌ [bd:advisor-k31] | ❌ [bd:advisor-2k3] | ❌ [bd:advisor-0k3] | ✅ (caretaking/caretaker loop) | ❌ [bd:advisor-87o] | ✅ in catalog | ❌ [bd:advisor-3y4] |
