+++ docs/wiki/index.json
@@ -205,8 +205,7 @@
       "Operator review gates dynamic skills due to prompt injection risk",
       "Advisor pattern layers Opus reviewer over Sonnet executor on review surfaces",
       "Advisor uses Claude Code subagent dispatch — never the Anthropic SDK directly",
-      "EpicMonitorLoop — stale epic detection and progress refresh",
-      "EpicSweeperLoop — auto-close completed epics",
+      "EpicMonitorLoop — stale epic detection, completion sweep, and progress refresh",
       "RetrospectiveLoop — durable-queue pattern analysis",
       "RunsGCLoop — artifact retention enforcement",
       "SecurityPatchLoop — Dependabot alert triage and issue filing",
