+++ src/cost_budget_watcher_loop.py
@@ -56,7 +56,6 @@ _TARGET_WORKERS = (
     "stale_issue_gc",
     "pr_unsticker",
     "epic_monitor",
-    "epic_sweeper",
     "principles_audit",
     "repo_wiki",
     "wiki_rot_detector",
