+++ src/epic.py
@@ -1062,6 +1062,84 @@ class EpicManager:
                     exc_info=True,
                 )
 
+    async def sweep_completed_epics(self) -> dict[str, int]:
+        """Auto-close open epics whose state/body sub-issue refs are all closed."""
+        epics = await self._fetcher.fetch_issues_by_labels(
+            self._config.epic_label, limit=50
+        )
+        if len(epics) == 50:
+            logger.warning(
+                "Epic monitor fetched exactly 50 epics for completion sweep — result "
+                "may be truncated; some epics may not be swept this cycle"
+            )
+
+        swept = 0
+        checked = 0
+        for epic in epics:
+            try:
+                sub_issues = self._collect_sweep_sub_issues(epic.number, epic.body)
+                if not sub_issues:
+                    continue
+                checked += 1
+                closed = await self._try_sweep_epic(epic.number, epic.body, sub_issues)
+                if closed:
+                    swept += 1
+            except Exception as exc:
+                reraise_on_credit_or_bug(exc)
+                logger.exception("Error sweeping epic #%d — skipping", epic.number)
+        return {"checked": checked, "swept": swept, "total_open_epics": len(epics)}
+
+    def _collect_sweep_sub_issues(self, epic_number: int, body: str) -> list[int]:
+        """Merge sub-issue refs from EpicState children and body checkboxes."""
+        refs: set[int] = set()
+
+        epic_state = self._state.get_epic_state(epic_number)
+        if epic_state is not None:
+            refs.update(epic_state.child_issues)
+
+        refs.update(parse_epic_sub_issues(body))
+        return sorted(refs)
+
+    async def _try_sweep_epic(
+        self, epic_number: int, epic_body: str, sub_issues: list[int]
+    ) -> bool:
+        """Close the epic if every referenced sub-issue is already closed."""
+        fixed_label = self._config.fixed_label[0] if self._config.fixed_label else ""
+
+        for issue_number in sub_issues:
+            issue = await self._fetcher.fetch_issue_by_number(issue_number)
+            if issue is None:
+                logger.warning(
+                    "Sub-issue #%d not found for epic #%d — skipping epic"
+                    " (remove stale ref from body to allow auto-close)",
+                    issue_number,
+                    epic_number,
+                )
+                return False
+            state_value = getattr(issue.state, "value", issue.state)
+            if state_value != GitHubIssueState.CLOSED.value:
+                return False
+
+        logger.info(
+            "All %d sub-issues closed for epic #%d — auto-closing",
+            len(sub_issues),
+            epic_number,
+        )
+
+        updated_body = check_all_checkboxes(epic_body)
+        if updated_body != epic_body:
+            await self._prs.update_issue_body(epic_number, updated_body)
+
+        if fixed_label:
+            await self._prs.add_labels(epic_number, [fixed_label])
+
+        comment = f"All {len(sub_issues)} sub-issues completed. Auto-closing epic."
+        await self._prs.post_comment(epic_number, comment)
+        await self._prs.close_issue(epic_number)
+        self._state.close_epic(epic_number)
+        self._invalidate_cache(epic_number)
+        return True
+
     async def trigger_release(self, epic_number: int) -> dict[str, object]:
         """Trigger async merge sequence and release creation for a bundled epic.
