+++ src/orchestrator.py
@@ -160,7 +160,6 @@ class HydraFlowOrchestrator:
             "merge_state_watcher": svc.merge_state_watcher_loop,
             "report_issue": svc.report_issue_loop,
             "epic_monitor": svc.epic_monitor_loop,
-            "epic_sweeper": svc.epic_sweeper_loop,
             "workspace_gc": svc.workspace_gc_loop,
             "runs_gc": svc.runs_gc_loop,
             "adr_reviewer": svc.adr_reviewer_loop,
@@ -984,7 +983,6 @@ class HydraFlowOrchestrator:
             ("merge_state_watcher", self._svc.merge_state_watcher_loop.run),
             ("report_issue", self._svc.report_issue_loop.run),
             ("epic_monitor", self._svc.epic_monitor_loop.run),
-            ("epic_sweeper", self._svc.epic_sweeper_loop.run),
             ("workspace_gc", self._svc.workspace_gc_loop.run),
             ("runs_gc", self._svc.runs_gc_loop.run),
             ("adr_reviewer", self._svc.adr_reviewer_loop.run),
