+++ tests/orchestrator_integration_utils.py
@@ -471,7 +471,6 @@ def build_scripted_services(
     services.report_issue_loop = FakeBackgroundLoop()
     services.epic_manager = MagicMock()
     services.epic_monitor_loop = FakeBackgroundLoop()
-    services.epic_sweeper_loop = FakeBackgroundLoop()
     services.workspace_gc_loop = FakeBackgroundLoop()
     services.runs_gc_loop = FakeBackgroundLoop()
     services.adr_reviewer_loop = FakeBackgroundLoop()
