+++ tests/regressions/test_issue_6735.py
@@ -1,16 +1,8 @@
 """Regression test for issue #6735.
 
-``EpicSweeperLoop._do_work()`` wraps each epic's sweep attempt in
-``except Exception: logger.exception(...)`` and continues.  This catches
-``AuthenticationError`` and ``CreditExhaustedError`` (both subclass
-``RuntimeError``) as if they were per-epic errors, then continues to the
-next item.  Auth failures and credit exhaustion should propagate to the
-background loop supervisor so the orchestrator can pause or stop, not be
-silently swallowed.
-
-These tests will be RED until the handler re-raises
-``AuthenticationError`` and ``CreditExhaustedError`` before the generic
-``except Exception`` block (e.g. via ``reraise_on_credit_or_bug``).
+Epic completion sweeps wrap each epic's sweep attempt in a generic exception
+handler so one transient failure does not abort the cycle. Auth failures and
+credit exhaustion still need to propagate to the background loop supervisor.
 """
 
 from __future__ import annotations
@@ -20,7 +12,6 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from epic_sweeper_loop import EpicSweeperLoop
 from subprocess_util import AuthenticationError, CreditExhaustedError
 from tests.conftest import IssueFactory
 from tests.helpers import make_bg_loop_deps
@@ -44,13 +35,14 @@ def _make_issue(
 
 
 def _make_epic(number: int, body: str):
-    """Create a mock epic issue with sub-issue refs in body."""
     return _make_issue(number, body=body, labels=["hydraflow-epic"])
 
 
 def _make_loop(tmp_path: Path):
-    """Build an EpicSweeperLoop with mock dependencies."""
-    deps = make_bg_loop_deps(tmp_path, enabled=True, epic_sweep_interval=3600)
+    from epic import EpicManager
+    from epic_monitor_loop import EpicMonitorLoop
+
+    deps = make_bg_loop_deps(tmp_path, enabled=True, epic_monitor_interval=60)
 
     fetcher = MagicMock()
     fetcher.fetch_issues_by_labels = AsyncMock(return_value=[])
@@ -64,12 +56,22 @@ def _make_loop(tmp_path: Path):
 
     state = MagicMock()
     state.get_epic_state = MagicMock(return_value=None)
+    state.close_epic = MagicMock()
 
-    loop = EpicSweeperLoop(
+    manager = EpicManager(
         config=deps.config,
-        fetcher=fetcher,
-        prs=prs,
         state=state,
+        prs=prs,
+        fetcher=fetcher,
+        event_bus=deps.bus,
+    )
+    manager.check_stale_epics = AsyncMock(return_value=[])
+    manager.refresh_cache = AsyncMock(return_value=None)
+    manager.get_all_progress = MagicMock(return_value=[])
+
+    loop = EpicMonitorLoop(
+        config=deps.config,
+        epic_manager=manager,
         deps=deps.loop_deps,
     )
     return loop, fetcher, prs, state
@@ -80,14 +82,9 @@ def _make_loop(tmp_path: Path):
 # ---------------------------------------------------------------------------
 
 
-class TestEpicSweeperPropagatesFatalErrors:
-    """AuthenticationError and CreditExhaustedError must propagate out of
-    ``_do_work``, not be caught by the per-epic handler."""
-
+class TestEpicMonitorPropagatesFatalErrors:
     @pytest.mark.asyncio
-    @pytest.mark.xfail(reason="Regression for issue #6735 — fix not yet landed", strict=False)
     async def test_authentication_error_propagates(self, tmp_path: Path) -> None:
-        """AuthenticationError must NOT be caught by except Exception."""
         loop, fetcher, _prs, _state = _make_loop(tmp_path)
         epic = _make_epic(100, "- [ ] #10\n- [ ] #20")
         fetcher.fetch_issues_by_labels = AsyncMock(return_value=[epic])
@@ -99,9 +96,7 @@ class TestEpicSweeperPropagatesFatalErrors:
             await loop._do_work()
 
     @pytest.mark.asyncio
-    @pytest.mark.xfail(reason="Regression for issue #6735 — fix not yet landed", strict=False)
     async def test_credit_exhausted_error_propagates(self, tmp_path: Path) -> None:
-        """CreditExhaustedError must NOT be caught by except Exception."""
         loop, fetcher, _prs, _state = _make_loop(tmp_path)
         epic = _make_epic(100, "- [ ] #10\n- [ ] #20")
         fetcher.fetch_issues_by_labels = AsyncMock(return_value=[epic])
@@ -114,7 +109,6 @@ class TestEpicSweeperPropagatesFatalErrors:
 
     @pytest.mark.asyncio
     async def test_plain_runtime_error_still_caught(self, tmp_path: Path) -> None:
-        """Plain RuntimeError should still be caught — loop continues."""
         loop, fetcher, _prs, _state = _make_loop(tmp_path)
         epic1 = _make_epic(100, "- [ ] #10")
         epic2 = _make_epic(200, "- [ ] #20")
