+++ tests/test_parameter_naming_consistency.py
@@ -113,7 +113,6 @@ class TestNoIssueNumVariables:
             "harness_insights",
             "epic",
             "changelog",
-            "epic_sweeper_loop",
             "orchestrator",
             "workspace_gc_loop",
             "timeline",
